switch to ES6-compatible minifier
Some downstream dependencies have introduced ES6 (bencode@1.0.0) and
there are others that have done so too, but which I've refrained from
updating (junk@2.0.0).

It's going to keep getting harder and harder to keep this code out of
the codebase -- so let's use babili, which is a minifier that supports
ES6.

It compresses worse than uglify (I think the bundle is something like
3KB larger now), but it's worth it IMO.