downgrade end-of-stream to v1.4.1
The minor update in our version range was breaking. We can update it later when it's fixed. 

even without an NPM bump, this is patch is useful to just let the CI tests work

issue: https://github.com/webtorrent/webtorrent/issues/1757 (should maybe keep that open until we can re-enable updates for the package though)